poa/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
httpx>=0.27.0
websockets>=12.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
jinja2>=3.1.0
jsonpath-ng>=1.6.0
jmespath>=1.0.0
jsonschema>=4.21.0
faker>=24.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.0.0
pyyaml>=6.0.0
apscheduler>=3.10.0
